add function description and remove K as input
First, by adding a description of the type:

    """
    Compute p-values from likelihood ratios.
    These are likelihood ratio test p-values.
    Parameters
    ----------
    null_lml : float
        Log of the marginal likelihood under the null hypothesis.
    alt_lmls : array_like
        Log of the marginal likelihoods under the alternative hypotheses.
    dof : int
        Degrees of freedom.
    Returns
    -------
    pvalues : ndarray
        P-values.
    """

when builing the read the docs, that becomes neat documentation for the same function
(as long as it is added to the index.rst file in the docs folder)

Second, I had left the option of providing K as an input to the CellRegMap runners
(run_association, run_interaction and estimate_betas).
The idea was that K (donors x donors) is more likely to be a file the user has.

However, CellRegMap takes cell-level vectors, so K would have to be expanded to become cells x cells
My function performed decomposition internally using Cholesky to go from K to hK.

However, that step needs to be done *prior* to expansion
i.e. K -> hK -> hK_expanded

we never want to build K_expanded, so the above steps need to be performed as
a preprocessing step, and hK_expanded needs to be inputed into the functions.

We are working on a wrapper function which can slice and dice and expand all files
including getting to the right SNPs for a given gene based on either a filter file
or a cis window, expanding G and K, loading expression from all genes and automating
parallelisation, but at this stage these steps have to be done by the user
(following tutorials from the website)